+++ driver/aws/dynamojournal/doc.go
@@ -1,0 +1,3 @@
+// Package dynamojournal provides an implementation of [journal.Store] that
+// persists to a DynamoDB table.
+package dynamojournal