+++ driver/aws/dynamojournal/store_test.go
@@ -1,26 +1,27 @@
-package dynamodb_test
+package dynamojournal_test
 
 import (
 	"context"
 	"testing"
 	"time"
 
-	. "github.com/dogmatiq/persistencekit/driver/aws/dynamodb"
+	. "github.com/dogmatiq/persistencekit/driver/aws/dynamojournal"
+	"github.com/dogmatiq/persistencekit/driver/aws/internal/dynamox"
 	"github.com/dogmatiq/persistencekit/journal"
 )
 
 func TestJournalStore(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 
-	client := newClient(t)
+	client := dynamox.NewTestClient(t)
 	table := "journal"
 
-	if err := CreateJournalStoreTable(ctx, client, table); err != nil {
+	if err := CreateTable(ctx, client, table); err != nil {
 		t.Fatal(err)
 	}
 
 	t.Cleanup(func() {
-		if err := deleteTable(ctx, client, table); err != nil {
+		if err := dynamox.DeleteTableIfNotExists(ctx, client, table); err != nil {
 			t.Fatal(err)
 		}
 
@@ -30,7 +31,7 @@ func TestJournalStore(t *testing.T) {
 	journal.RunTests(
 		t,
 		func(t *testing.T) journal.Store {
-			return &JournalStore{
+			return &Store{
 				Client: client,
 				Table:  table,
 			}