+++ driver/aws/dynamokv/doc.go
@@ -1,0 +1,3 @@
+// Package dynamokv provides a [kv.Store] implementation that persists to
+// a DynamoDB table.
+package dynamokv