+++ driver/aws/dynamokv/table.go
@@ -1,0 +1,68 @@
+package dynamokv
+
+import (
+	"context"
+	"errors"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
+	"github.com/dogmatiq/persistencekit/driver/aws/internal/awsx"
+)
+
+const (
+	keyspaceAttr = "Keyspace"
+	keyAttr      = "Key"
+	valueAttr    = "Value"
+)
+
+// CreateTable creates a DynamoDB table for use with [Store].
+func CreateTable(
+	ctx context.Context,
+	client *dynamodb.Client,
+	table string,
+	decorators ...func(*dynamodb.CreateTableInput) []func(*dynamodb.Options),
+) error {
+	_, err := awsx.Do(
+		ctx,
+		client.CreateTable,
+		func(in *dynamodb.CreateTableInput) []func(*dynamodb.Options) {
+			var options []func(*dynamodb.Options)
+			for _, dec := range decorators {
+				options = append(options, dec(in)...)
+			}
+
+			return options
+		},
+		&dynamodb.CreateTableInput{
+			TableName: aws.String(table),
+			AttributeDefinitions: []types.AttributeDefinition{
+				{
+					AttributeName: aws.String(keyspaceAttr),
+					AttributeType: types.ScalarAttributeTypeS,
+				},
+				{
+					AttributeName: aws.String(keyAttr),
+					AttributeType: types.ScalarAttributeTypeB,
+				},
+			},
+			KeySchema: []types.KeySchemaElement{
+				{
+					AttributeName: aws.String(keyspaceAttr),
+					KeyType:       types.KeyTypeHash,
+				},
+				{
+					AttributeName: aws.String(keyAttr),
+					KeyType:       types.KeyTypeRange,
+				},
+			},
+			BillingMode: types.BillingModePayPerRequest,
+		},
+	)
+
+	if errors.As(err, new(*types.ResourceInUseException)) {
+		return nil
+	}
+
+	return err
+}