+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 - **[BC]** Moved `postgres.KeyValueStore` to `pgkv.Store`
 - **[BC]** Moved `memory.JournalStore` to `memoryjournal.Store`
 - **[BC]** Moved `memory.KeyValueStore` to `memorykv.Store`
+- **[BC]** Moved `dynamodb.JournalStore` to `dynamojournal.Store`
+- **[BC]** Moved `dynamodb.KeyValueStore` to `dynamokv.Store`
 
 ## [0.1.0] - 2023-10-16